First commit to the project
Created a Main class that prompts users for either account or login and runs all of the other methods for the project. Created a Key-Value Map with the key as the username and the value is the account information. This is so that when the user logs in, they will pair the username and check if the password is correct.
Created a CreateAccount class that prompts the user for information such as first name, last name, email, username, and password.
Created a Accounts class that has the users information such as first name, last name, email, username, and password.